requirements.txt
streamlit>=1.57.0
langchain>=1.2.0
langchain-community>=0.4.0
langchain-text-splitters>=0.3.0
pypdf>=6.0.0
sentence-transformers>=3.0.0
pinecone>=5.1.0
python-dotenv>=1.0.0
requests>=2.32.0
google-genai>=1.7.0
grpcio>=1.60.0
googleapis-common-protos>=1.63.0